+++ backend/canisters/user/api/src/updates/c2c_send_messages.rs
@@ -28,6 +28,7 @@ pub enum Response {
 pub enum C2CReplyContext {
     ThisChat(MessageId),
     OtherChat(ChatId, EventIndex),
+    OtherEventList(ChatId, Option<MessageIndex>, EventIndex),
 }
 
 impl Args {