+++ backend/canisters/user/impl/src/updates/send_message.rs
@@ -127,6 +127,9 @@ fn validate_request(args: &Args, runtime_state: &RuntimeState) -> ValidateReques
             ContentValidationError::UnauthorizedToSendProposalMessages => {
                 InvalidRequest("User unauthorized to send proposal messages".to_string())
             }
+            ContentValidationError::Unauthorized => {
+                InvalidRequest("User unauthorized to send messages of this type".to_string())
+            }
         })
     } else if args.recipient == my_user_id {
         ValidateRequestResult::Valid(UserType::_Self)
@@ -171,6 +174,7 @@ fn send_message_impl(
             sender_message_index: message_event.event.message_index,
             content: args.content.into(),
             replies_to: args.replies_to.and_then(|r| {
+                // TODO switch to `event_list_if_other` once all user canisters are upgraded
                 if let Some(chat_id) = r.chat_id_if_other {
                     Some(C2CReplyContext::OtherChat(chat_id, r.event_index))
                 } else {