+++ frontend/openchat-agent/src/services/userIndex/candid/types.d.ts
@@ -55,7 +55,8 @@ export interface CanisterWasm {
 export interface Challenge { 'key' : ChallengeKey, 'png_base64' : string }
 export interface ChallengeAttempt { 'key' : ChallengeKey, 'chars' : string }
 export type ChallengeKey = number;
-export type ChatEvent = { 'MessageReactionRemoved' : UpdatedMessage } |
+export type ChatEvent = { 'Empty' : null } |
+  { 'MessageReactionRemoved' : UpdatedMessage } |
   { 'ParticipantJoined' : ParticipantJoined } |
   { 'ParticipantAssumesSuperAdmin' : ParticipantAssumesSuperAdmin } |
   { 'GroupDescriptionChanged' : GroupDescriptionChanged } |
@@ -114,6 +115,7 @@ export interface ChatMetrics {
   'file_messages' : bigint,
   'poll_votes' : bigint,
   'text_messages' : bigint,
+  'message_reminders' : bigint,
   'image_messages' : bigint,
   'replies' : bigint,
   'video_messages' : bigint,
@@ -534,7 +536,9 @@ export type MessageContent = { 'Giphy' : GiphyContent } |
   { 'Audio' : AudioContent } |
   { 'Crypto' : CryptoContent } |
   { 'Video' : VideoContent } |
-  { 'Deleted' : DeletedContent };
+  { 'Deleted' : DeletedContent } |
+  { 'MessageReminderCreated' : MessageReminderCreated } |
+  { 'MessageReminder' : MessageReminder };
 export type MessageContentInitial = { 'Giphy' : GiphyContent } |
   { 'File' : FileContent } |
   { 'Poll' : PollContent } |
@@ -545,7 +549,9 @@ export type MessageContentInitial = { 'Giphy' : GiphyContent } |
   { 'Audio' : AudioContent } |
   { 'Crypto' : CryptoContent } |
   { 'Video' : VideoContent } |
-  { 'Deleted' : DeletedContent };
+  { 'Deleted' : DeletedContent } |
+  { 'MessageReminderCreated' : MessageReminderCreated } |
+  { 'MessageReminder' : MessageReminder };
 export interface MessageEventWrapper {
   'event' : Message,
   'timestamp' : TimestampMillis,
@@ -569,6 +575,15 @@ export interface MessageMatch {
 export interface MessagePinned {
   'pinned_by' : UserId,
   'message_index' : MessageIndex,
+}
+export interface MessageReminder {
+  'notes' : [] | [string],
+  'reminder_id' : bigint,
+}
+export interface MessageReminderCreated {
+  'notes' : [] | [string],
+  'remind_at' : TimestampMillis,
+  'reminder_id' : bigint,
 }
 export interface MessageUnpinned {
   'due_to_message_deleted' : boolean,
@@ -817,6 +832,7 @@ export type RemovePlatformModeratorResponse = { 'Success' : null } |
 export interface RemovePlatformOperatorArgs { 'user_id' : UserId }
 export type RemovePlatformOperatorResponse = { 'Success' : null };
 export interface ReplyContext {
+  'event_list_if_other' : [] | [[ChatId, [] | [MessageIndex]]],
   'chat_id_if_other' : [] | [ChatId],
   'event_index' : EventIndex,
 }