+++ frontend/openchat-client/src/openchat.ts
@@ -3797,6 +3797,12 @@ export class OpenChat extends EventTarget {
         return this.api.updateMarketMakerConfig(config);
     }
 
+    usernameAndIcon(user?: PartialUserSummary): string {
+        return user !== undefined 
+            ? `${user?.username}  ${user?.diamond ? "💎" : ""}` 
+            : this.config.i18nFormatter("unknownUser");
+    }    
+
     diamondDurationToMs = diamondDurationToMs;
 
     /**