+++ frontend/openchat-shared/src/utils/chat.ts
@@ -96,6 +96,7 @@ export function userIdsFromEvents(events: EventWrapper<ChatEvent>[]): Set<string
             case "aggregate_common_events":
             case "chat_frozen":
             case "chat_unfrozen":
+            case "empty":
                 break;
             default:
                 throw new UnsupportedValueError("Unexpected ChatEvent type received", e.event);