+++ backend/canisters/group/CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Implement 'Gated Groups' ([#3406](https://github.com/open-ic/open-chat/pull/3406))
 - Added `register_proposal_vote_v2` for voting on proposals directly from the frontend ([#3413](https://github.com/open-ic/open-chat/pull/3413))
+- Added `Empty` event type ([#3439](https://github.com/open-ic/open-chat/pull/3439))
+- Added new message content types for reminders ([#3440](https://github.com/open-ic/open-chat/pull/3440))
 
 ### Removed